+++ tasks.py
@@ -127,8 +127,8 @@ def release_github(ctx):
 
 @task
 def commit(ctx):
-    ctx.run("git commit -a -m \"v%s release\"" % ver)
-    ctx.run("git push")
+    ctx.run("git commit -a -m \"v%s release\"" % ver, warn=True)
+    ctx.run("git push", warn=True)
 
 
 @task